+++ tests/protocols/streaming.rs
@@ -1,0 +1,4 @@
+use playground::*;
+
+// Recursive/streaming protocol
+pub type Streaming = TRec<Http, TInteract<Http, TClient, Message, TEnd<Http>>>;