+++ tests/trybuild/duplicate_roles_par.rs
@@ -1,0 +1,7 @@
+use playground::*;
+
+type DupRolePar = tpar!(Http;
+    TInteract<Http, TClient, Message, TEnd<Http>>,
+    TInteract<Http, TClient, Publish, TEnd<Http>>
+);
+assert_disjoint!(par DupRolePar);