+++ tests/trybuild/empty_choice.rs
@@ -1,0 +1,3 @@
+use playground::*;
+
+type EmptyChoice = tchoice!(Http;);