+++ tests/trybuild/empty_par.rs
@@ -1,0 +1,3 @@
+use playground::*;
+
+type EmptyPar = tpar!(Http;);