+++ tests/trybuild/trailing_commas.rs
@@ -1,0 +1,11 @@
+use playground::*;
+
+type TrailingCommaChoice = tchoice!(Http;
+    TInteract<Http, TClient, Message, TEnd<Http>>,
+    TInteract<Http, TServer, Response, TEnd<Http>>,
+);
+
+type TrailingCommaPar = tpar!(Http;
+    TInteract<Http, TClient, Message, TEnd<Http>>,
+    TInteract<Http, TServer, Response, TEnd<Http>>,
+);